+++ CHANGELOG.md
@@ -6,3 +6,4 @@ Initial version. Roughly, this is a spin-off of the "stool build" command from
 Stool 6 https://github.com/mlhartme/stool, but
 
 * build arguments no longer added as labels
+* switch from stool docker client to https://github.com/docker-java/docker-java